limits library included for computing max weight of cost function

+++ inc/heuristic.h
@@ -7,8 +7,8 @@
 	#include "D:\Projects\PRGM_WORKCRAFT\inc\config.h"
 #endif
 
-float area_encodings_ssd(int, int, float*,int,int);
-float weight_function(int,int);
+long long int area_encodings_ssd(int, int, long long int*,int,int);
+long long int weight_function(int,int);
 void HD_min_v2(int*,int,int*,int*,int,int,int*,int,int,int);
 void HD_min_v3(int*,int,int*,int*,int,int,int*,int,int,int);
 void random_opcode_choice_v2(int*,int,int*,int*,int,int,int*,int,int,int);
