+++ inc/prgm_helper.h
@@ -9,7 +9,7 @@
 
 int parse_arg(int,char**);
 void print_help(char*);
-int manage_data_base(int,float,int,float,int,int*);
+int manage_data_base(int,long long int,int,long long int,int,int*);
 int read_set_encoding(int, int*);
 void removeTempFiles();
 void print_version();
